# [IMP] Improve performance on sql